requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
decorator==4.4.2
future==0.18.2
idna==2.10
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==12.7
pytz==2021.1
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
